Re-register scheduled commands
